Extract all admin commands to their own script files.
