+++ admin_commands/addFeat.js
@@ -1,8 +1,8 @@
 'use strict';
 
 const _         = require('../src/helpers.js');
+const util      = require('util');
 const { Feats } = require('./feats');
-
 
 exports.command = (rooms, items, players, npcs, Commands) =>
       (args, player) => {
