+++ admin_commands/character.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const util = require('util');
+
 exports.command = (rooms, items, players, npcs, Commands) =>
   (args, player) => {
     const attrs = player.getAttributes();
