+++ src/commands.js
@@ -97,125 +97,7 @@ const Commands = {
     }),
   },
 
-  admin_commands: {
-
-      debugInv: (rooms, items, players, npcs, Commands) =>
-        (args, player) => {
-          const inv = player.getInventory();
-          player.warn("ITEMS:\n");
-          for (let i in inv) {
-            const item = inv[i];
-            player.say(item.getShortDesc());
-
-            const attrs = item.getAttributes();
-            for (let attr in attrs) {
-              player.say(attr + ': ' + attrs[attr]);
-            }
-
-            const prereqs = item.getPrerequisites();
-            for (let prereq in prereqs) {
-              player.say(prereq + ': ' + prereqs[prereq]);
-            }
-
-            const isContainer = item.isContainer();
-            const actualWeight = item.getWeight(items);
-            player.say(`Is container: ${isContainer}`);
-            player.say(`Total weight: ${actualWeight}`);
-            if (isContainer) {
-              const itemContents   = item.getInventory();
-              const spaceLeft      = item.getRemainingSizeCapacity(items);
-              const contentsWeight = item.getContainerWeight(items);
-             
-              player.say(`Contents: ${itemContents}`);
-              player.say(`Space left: ${spaceLeft}`);
-              player.say(`Contents weight: ${contentsWeight}`);
-              player.say(`
-              ===========`);
-            }
-
-            player.say(item.isEquipped() ? 'Equipped' : 'In inventory');
-            player.say(`========`);
-            player.say(`Behaviors: ${item.behaviors}, script: ${item.script}`);
-            const events = item.eventNames();
-            player.say(`Events: ${events.join(', ')}`);
-            player.warn('========\n');
-            console.log('events for ', item.getShortDesc());
-            console.log(events);
-          }
-
-        },
-
-    debugItems: (rooms, items, players, npcs, Commands) =>
-      (args, player) => {
-        const allItems = items.objects;
-        player.warn(`UUIDs: ${Object.keys(allItems)}`);
-        player.warn(`========`);
-        player.warn("GLOBAL ITEMS:\n");
-        for (let uid in allItems) {
-          const item = allItems[uid];
-
-          player.say(item.getShortDesc());
-          player.say('\n');
-          player.say('vnum: ' + item.getVnum());
-          player.say('uuid: ' + item.getUuid());
-          player.say('location: ' + item.getRoom());
-          player.say('\n')
-          player.say(item.isEquipped() ? 'Equipped' : 'Not equipped');
-
-          const container = item.getContainer();
-          player.say(container ? 'Container: ' + container : 'No container.');
-
-          const inventory = item.getInventory();
-          player.say(inventory ? 'Inventory: ' : 'No inventory.');
-          if (inventory) {
-            for (let i in inventory) {
-              player.say(`${i}: ${inventory[i].getShortDesc()} ${i.getUuid()}`);
-            }
-          }
-          player.warn('========\n');
-        }
-
-      },
-
-    setAttribute: (rooms, items, players, npcs, Commands) =>
-      (args, player) => {
-        args = _.splitArgs(args);
-
-        const attributes = player.getAttributes();
-        const attr = args[0];
-
-        if (attr in attributes) {
-          const score = parseInt(args[1], 10);
-          if (!score || isNaN(score)) {
-            return player.say('<red>ADMIN: Not a real number.</red>');
-          }
-
-          player.setAttribute(attr, score);
-          util.log("@@Admin: " + player.getName() + " set attr " + attr + " to " + score + ".");
-          return player.say("<red>ADMIN: Set " + attr + " to " + score + ".</red>");
-        }
-
-        player.say('<red>ADMIN: No such attribute.</red>');
-      },
-
-    teleport: (rooms, items, players, npcs, Commands) =>
-      (args, player) => {
-        if (!player || !player.say || !args) { return; }
-        const vnum = parseInt(args, 10);
-        if (isNaN(vnum)) {
-          return player.say("<red>ADMIN: Invalid vnum.</red>");
-        }
-
-        if (rooms.getAt(vnum)) {
-          player.setLocation(vnum);
-          player.say("<red>ADMIN: You have teleported.");
-          return Commands.player_commands.look(null, player);
-        }
-
-        player.say("<red>ADMIN: 404: Room not found.</red>");
-
-      }
-  },
+  admin_commands: {},
 
   /**
    * Configure the commands by using a joint players/rooms array
